v2g2t/requirements.txt
azure-functions==1.2.0
ffmpeg-python==0.2.0